style: removes circular imports from the library
This library is being detected as having circular imports which can prevent its use in some
projects. This changes all the `import {<thing>} from '..'` lines to instead import directly from
the files that have the required dependencies. This also moves some types to their own file to
remove circular imports from the `utils` directory. Otherwise this change makes no functional change
to any hooks.

905
