+++ src/useCustomCompareMemo/useCustomCompareMemo.ts
@@ -1,7 +1,7 @@
 import { useMemo, useRef } from 'react';
 
 import type { DependencyList } from 'react';
-import type { DependenciesComparator } from '..';
+import type { DependenciesComparator } from '../types';
 
 /**
  * Like useMemo but uses provided comparator function to validate dependency changes.
